children and parent on obras implemented

+++ src/services/obra/DeleteObraService.ts
@@ -10,7 +10,7 @@ export class DeleteObraService {
         
         const repo = getRepository(Obra);
         const obra = await repo.findOne(id, {
-            relations: ["bibliografias", "obrasAutores"]
+            relations: ["bibliografias", "obrasAutores", "obras"]
         });
         
         if(!obra){
@@ -23,6 +23,10 @@ export class DeleteObraService {
 
         if(obra.obrasAutores.length > 0){
             return new Error("Obra com autores cadastrados");
+        }
+
+        if(obra.obrasChildren.length > 0){
+            return new Error("Obra com obras cadastradas");
         }
 
         await repo.delete(id);
